pyproject.toml
[project]
name = "test-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "blockfrost-python==0.6.0",
    "cryptography==46.0.3",
    "fastapi==0.116.1",
    "itsdangerous>=2.2.0",
    "pyjwt==2.8.0",
    "psycopg2-binary>=2.9.11",
    "pycardano==0.17.0",
    "pydantic==2.11.7",
    "pydantic-settings==2.12.0",
    "pytest==9.0.1",
    "python-dotenv==1.1.1",
    "redis==6.4.0",
    "requests==2.32.3",
    "sqlalchemy==2.0.27",
    "starlette==0.47.3",
    "uvicorn==0.35.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]
